requirements.txt
Django==5.1.*
djangorestframework==3.15.*
django-filter==24.2
psycopg[binary]==3.2.*
django-environ==0.11.*
celery==5.4.*
redis==5.0.*
openpyxl==3.1.*
icalendar==5.0.*
pytest==8.3.*
pytest-django==4.9.*
ruff==0.6.*
black==24.8.*
